Refactor UI code into modules in `lib-ui`
It's starting to git unwieldy!
